flask-memo-app/requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
pymysql
cryptography
Flask-Login